Remove dependency of BundleContext from Aggregator Implementation.
